Add support for dynamically updated facets
Some facet interfaces are enriched by updating the facets after each
new Ajax request is performed.  This commit adds support for this.

The key is to use Event Delegation (learn.jquery.com/events/event-delegation)
to bind the listener to an unchanging parent element (the form itself) rather
than directly to the inputs.  This way, inputs can change with the Ajax
requests, and the binding is not broken.

Implementation requires a custom postAJAX function to be specified in the
plugin configuration, and for the updated facets to be returned in some format
that can be processed by that function.